IssueNo:#IAHA4A Description:fix codecheck Sig:SIG_ApplicaitonFramework Feature or Bugfix:Bugfix Binary Source:No Signed-off-by: wangtiantian <wangtiantian19@huawei.com>

+++ services/bundlemgr/src/bundle_data_storage_rdb.cpp
@@ -21,7 +21,7 @@ namespace OHOS {
 namespace AppExecFwk {
 namespace {
 constexpr const char* BUNDLE_RDB_TABLE_NAME = "installed_bundle";
-const int32_t CLOSE_TIME = 15; // delay 15s stop rdbStore
+const int8_t CLOSE_TIME = 15; // delay 15s to backup
 }
 BundleDataStorageRdb::BundleDataStorageRdb()
 {
